Time
Some viewers have questioned and speculated the possibility of a deeper
meaning behind the video, leading to critical analyses of the lyrics
that are both serious and facetious. One of the leading theories
interprets the film as a metaphor to how children are conditioned to act
in a certain way due to some of the lines spoken by the notebook.